improving validations

+++ src/app/controllers/ContactController.js
@@ -1,5 +1,6 @@
-const { response } = require('express');
-const ContactsRepository = require('../repositories/ContactsRepository');
+const { response } = require("express");
+const ContactsRepository = require("../repositories/ContactsRepository");
+const isValidUUID = require("../utils/isValidUUID");
 
 class ContactController {
     async index(req, res) {
@@ -12,11 +13,15 @@ class ContactController {
     async show(req, res) {
         // List one register
         const { id } = req.params;
+
+        if (!isValidUUID(id)) {
+            return res.status(400).json({ error: "The ID is incorrect." });
+        }
+
         const contact = await ContactsRepository.findById(id);
         if (!contact) {
-            return res.status(404).json({ error: 'User Not Found' });
+            return res.status(404).json({ error: "User Not Found" });
         }
-        console.log(contact);
         res.json(contact);
     }
 
@@ -24,22 +29,30 @@ class ContactController {
         // Create new register
         const { name, email, phone, category_id } = req.body;
 
-        if (!name) {
-            return res.status(400).json({ error: 'Name is required' });
+        if (category_id && !isValidUUID(category_id)) {
+            return res.status(400).json({ error: "Invalid category" });
         }
 
-        const contactExists = await ContactsRepository.findByEmail(email);
+        if (!name) {
+            return res.status(400).json({ error: "Name is required" });
+        }
 
-        if (contactExists){
-            // 400: Bad Request
-            return res.status(400).json({ error: 'This email already exists' });
+        if (email) {
+            const contactExists = await ContactsRepository.findByEmail(email);
+
+            if (contactExists) {
+                // 400: Bad Request
+                return res
+                    .status(400)
+                    .json({ error: "This email already exists" });
+            }
         }
 
         const contact = await ContactsRepository.create({
             name,
-            email,
+            email: email || null,
             phone,
-            category_id
+            category_id: category_id || null,
         });
 
         res.json(contact);
@@ -50,35 +63,48 @@ class ContactController {
         const { id } = req.params;
         const { name, email, phone, category_id } = req.body;
 
+        if (!isValidUUID(id)) {
+            return res.status(400).json({ error: "The ID is incorrect." });
+        }
+
+        if (category_id && !isValidUUID(category_id)) {
+            return res.status(400).json({ error: "Invalid category" });
+        }
+
+        if (!name) {
+            return res.status(400).json({ error: "name is required" });
+        }
         const contactExists = await ContactsRepository.findById(id);
-        const contactByEmail = await ContactsRepository.findByEmail(email);
-        console.log(contactByEmail);
         if (!contactExists) {
             return res.status(400).json({ error: "This user does't exists" });
         }
-        if (!name) {
-            return res.status(400).json({ error: 'name is required' });
-        }
-        if (contactByEmail && contactByEmail.id !== id) {
-            return res.status(400).json({ error: 'email already exists' });
+
+        if (email) {
+            const contactByEmail = await ContactsRepository.findByEmail(email);
+
+            if (contactByEmail && contactByEmail.id !== id) {
+                return res.status(400).json({ error: "email already exists" });
+            }
         }
 
         const contact = await ContactsRepository.update({
             id,
             name,
-            email,
+            email: email || null,
             phone,
-            category_id
+            category_id: category_id || null,
         });
 
         res.status(200).json(contact);
-
-
     }
 
     async delete(req, res) {
         // Delete one register
         const { id } = req.params;
+
+        if (!isValidUUID(id)) {
+            return res.status(400).json({ error: "The ID is incorrect." });
+        }
 
         await ContactsRepository.delete(id);
 
@@ -87,4 +113,4 @@ class ContactController {
     }
 }
 
-module.exports = new ContactController;
+module.exports = new ContactController();
